3 course
I tried beginning to code the product page, but I ended up only thinking how. I did however do the CSS for the size images, the colours of the backgrounds with layout (smacss) and lastly I also added the h2 size. Last time I tried doing the link anchor with the nav and the corresponding categories but it didnt work, this didnt change today.