I am blocked
I tried doing the planed CSS but I got lost on how to do it. Flexbox? Bootstrap? Media queries? I listened and payed attention in class but when it comes to the practical part I get confused. I hope that I will learn and find a other way to do things with this break.